moved log.cpp and log.hpp to core and added engine logs to validation
